1st Commit
This is a primitive cannon.

The cannon itself works pretty well. The cannon ball physics are not
satisfactory.

Note that I did not use the built in physics functions to achieve this.
I used physics and mathematics to achieve  what I did.

What I should have done:
1) Comment my code (please forgive me on this, it's late).
2) Added proper physics to my cannon ball. All it does is have a
projectile motion and fall to the ground without any bounce up.

Hoping to do:
1) Comment my code
2) Re-test using actual physics